Split `signal_tools` into submodules and implement lazy loading
