+++ hyperspy/signal_tools/message.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+# Copyright 2007-2025 The HyperSpy developers
+#
+# This file is part of HyperSpy.
+#
+# HyperSpy is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# HyperSpy is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with HyperSpy. If not, see <https://www.gnu.org/licenses/#GPL>.
+
+import traits.api as t
+
+from hyperspy.ui_registry import add_gui_method
+
+
+@add_gui_method(toolkey="hyperspy.SimpleMessage")
+class SimpleMessage(t.HasTraits):
+    text = t.Str
+
+    def __init__(self, text=""):
+        self.text = text
